move asset environment stuff to separate file.
