+++ amable/__init__.py
@@ -6,10 +6,6 @@ from dotenv import load_dotenv
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 
-from flask_assets import Environment, Bundle
-from sass import compile as sass_compile
-
-from amable.blueprints.base import base
 
 load_dotenv(join(dirname(__file__), '..', '.env'))
 
@@ -17,42 +13,9 @@ app = Flask(__name__)
 app.config.from_envvar('AMABLE_SETTINGS')
 
 
-def sass_filter(_in, out, **kw):
-    compiled_sass = sass_compile(
-        string=_in.read(),
-        include_paths=[join(dirname(__file__), 'assets', 'css', 'lib')]
-    )
+from amable.blueprints.base import base
 
-    out.write(compiled_sass)
-
-
-env = Environment(app)
-
-env.load_path = [
-    join(dirname(__file__), 'assets', 'css'),
-    join(dirname(__file__), 'assets', 'jsc')
-]
-
-
-env.register(
-    'css',
-    Bundle(
-        'application.scss',
-        filters=[sass_filter],
-        output='application.css'
-    )
-)
-
-env.register(
-    'jsc',
-    Bundle(
-        'lib/jquery-3.1.1.min.js',
-        'application.js',
-        filters=['jsmin'],
-        output='application.js'
-    )
-)
-
+from amable.utils.assets import assets_env
 
 from models import db
 from models.user import User
